requirements.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bcrypt==4.2.1
beanie==1.29.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
sentry-sdk[fastapi]==2.54.0
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
lazy-model==0.2.0
motor==3.7.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
redis==5.2.1
requests==2.28.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
openai>=1.0.0
APScheduler==3.10.4